feat(logger): update pino
