+++ nestjs/package.json
@@ -38,14 +38,14 @@
     "cookie-parser": "1.4.6",
     "jsonwebtoken": "8.5.1",
     "lodash": "4.17.21",
-    "nestjs-pino": "2.4.0",
+    "nestjs-pino": "2.5.0",
     "passport": "0.5.2",
     "passport-custom": "1.1.1",
     "passport-github2": "0.1.12",
     "passport-http": "0.3.0",
     "passport-jwt": "4.0.0",
     "pg": "8.7.1",
-    "pino-http": "6.4.0",
+    "pino-http": "6.5.0",
     "reflect-metadata": "0.1.13",
     "rimraf": "3.0.2",
     "rxjs": "7.5.2",
